requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.1
django-cors-headers==2.4.0
django-debug-toolbar==1.10.1
django-filter==2.0.0
djangorestframework==3.8.2
drf-yasg==1.10.2
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonschema==2.6.0
Markdown==2.6.11
MarkupSafe==1.0
mutagen==1.41.1
Pillow==5.2.0
pkg-resources==0.0.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
ruamel.yaml==0.15.70
six==1.11.0
sqlparse==0.2.4
swagger-spec-validator==2.4.0
uritemplate==3.0.0
urllib3==1.23
